Code move and linter
