+++ .github/configs/wordlist.txt
@@ -749,3 +749,11 @@ Cvetkovic
 cvetkovic
 ethz
 lazar
+CooldownSeconds
+Dirigent
+RpsColdStartRatioPercentage
+RpsIterationMultiplier
+RpsMemoryMB
+RpsRuntimeMs
+RpsTarget
+SQRT
