+++ cmd/loader.go
@@ -30,7 +30,6 @@ import (
 	"fmt"
 	"github.com/vhive-serverless/loader/pkg/generator"
 	"golang.org/x/exp/slices"
-	"math/rand"
 	"os"
 	"strconv"
 	"strings"
@@ -255,69 +254,8 @@ func runRPSMode(cfg *config.LoaderConfiguration) {
 		LoaderConfiguration: cfg,
 		TraceDuration:       determineDurationToParse(cfg.ExperimentDuration, cfg.WarmupDuration),
 
-		Functions: createRPSFunctions(cfg, warmFunction, warmStartCount, coldFunctions, coldStartCount),
+		Functions: generator.CreateRPSFunctions(cfg, warmFunction, warmStartCount, coldFunctions, coldStartCount),
 	})
 
 	experimentDriver.RunExperiment(true, false)
-}
-
-func createRPSFunctions(cfg *config.LoaderConfiguration, warmFunction common.IATArray, warmFunctionCount []int,
-	coldFunctions []common.IATArray, coldFunctionCount [][]int) []*common.Function {
-	var result []*common.Function
-
-	result = append(result, &common.Function{
-		Name: fmt.Sprintf("warm-function-%d", rand.Int()),
-
-		InvocationStats: &common.FunctionInvocationStats{Invocations: warmFunctionCount},
-		MemoryStats:     &common.FunctionMemoryStats{Percentile100: float64(cfg.RpsMemoryMB)},
-		DirigentMetadata: &common.DirigentMetadata{
-			Image:               "trace",
-			Port:                80,
-			Protocol:            "tcp",
-			ScalingUpperBound:   1024,
-			ScalingLowerBound:   1,
-			IterationMultiplier: cfg.RpsIterationMultiplier,
-		},
-
-		Specification: &common.FunctionSpecification{
-			IAT:                  warmFunction,
-			RuntimeSpecification: createRuntimeSpecification(len(warmFunction), cfg.RpsRuntimeMs, cfg.RpsMemoryMB),
-		},
-	})
-
-	for i := 0; i < len(coldFunctions); i++ {
-		result = append(result, &common.Function{
-			Name: fmt.Sprintf("cold-function-%d-%d", i, rand.Int()),
-
-			InvocationStats: &common.FunctionInvocationStats{Invocations: coldFunctionCount[i]},
-			MemoryStats:     &common.FunctionMemoryStats{Percentile100: float64(cfg.RpsMemoryMB)},
-			DirigentMetadata: &common.DirigentMetadata{
-				Image:               "trace",
-				Port:                80,
-				Protocol:            "tcp",
-				ScalingUpperBound:   1,
-				ScalingLowerBound:   0,
-				IterationMultiplier: cfg.RpsIterationMultiplier,
-			},
-
-			Specification: &common.FunctionSpecification{
-				IAT:                  coldFunctions[i],
-				RuntimeSpecification: createRuntimeSpecification(len(coldFunctions[i]), cfg.RpsRuntimeMs, cfg.RpsMemoryMB),
-			},
-		})
-	}
-
-	return result
-}
-
-func createRuntimeSpecification(count int, runtime, memory int) common.RuntimeSpecificationArray {
-	var result common.RuntimeSpecificationArray
-	for i := 0; i < count; i++ {
-		result = append(result, common.RuntimeSpecification{
-			Runtime: runtime,
-			Memory:  memory,
-		})
-	}
-
-	return result
 }
